Added shortcuts for directories

+++ set.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 # Add to ~/.bashrc:
-#   source "/home/$USER/Soft/fastcd/set.sh"
+#   source "$HOME/Soft/fastcd/set.sh"
 # Do not forget to install extra modules: pip install --user urwid psutil
 
 FASTCDTOOLS="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
@@ -13,7 +13,7 @@ python $REFRESHERTOOL --daemon
 function j {
     PATHFILE="/tmp/`date +%s`.path"
     python $JUMPERTOOL --escape-special-symbols -o $PATHFILE $@
-    if [[ ! $@ == "--help" ]] && [[ ! $@ == "-h" ]]
+    if [[ ! $@ == "--help" ]] && [[ ! $@ == "-h" ]] && [[ ! $@ == "--list-shortcut-paths" ]] && [[ ! $@ == "-l" ]]
     then
         OUTPUTPATH=`cat $PATHFILE`
         rm $PATHFILE
